Sound filter
This is the sound filter part of my Jump! game which greatly increase the accuracy of getting player's sound from ambient sound. I first make a volumeThreshold which make the little ambient sound to be ignored. Then I collect 1024 effect sample effect volume and get the highest one to return. After 6 such highest volume picked up, I picked the fifth volume to be the real effect sound and use that to be the factor time the jump force.